fixing by separating out NestedRow
